ref: Reorganize Day 2 files
