+++ lib/puzzles/2023/day02/input.txt
