+++ test/puzzles/2023/day02/day02_test.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require "test_helper"
-require "puzzles/2023/day02"
+require "puzzles/2023/day02/day02"
 
 module AdventOfCode
   module Test
@@ -33,8 +33,7 @@ module AdventOfCode
           end
 
           def test_answer_input_set
-            input_file = "#{File.dirname(__FILE__)}/input.txt"
-            puzzle = AdventOfCode::Puzzles2023::Day02::Part1.new(file: input_file)
+            puzzle = AdventOfCode::Puzzles2023::Day02::Part1.new
 
             assert_equal 2278, puzzle.answer(blue: 14, green: 13, red: 12)
           end
@@ -66,8 +65,7 @@ module AdventOfCode
           end
 
           def test_answer_input_set
-            input_file = "#{File.dirname(__FILE__)}/input.txt"
-            puzzle = AdventOfCode::Puzzles2023::Day02::Part2.new(file: input_file)
+            puzzle = AdventOfCode::Puzzles2023::Day02::Part2.new
 
             assert_equal 67_953, puzzle.answer
           end
